Moved shared definitions.
